# lib/api-zod/src/generated/types/analysisInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FarmInput } from './farmInput';
import type { Weather } from './weather';

export type AnalysisInput = FarmInput & ({
  /** @nullable */
  farmProfileId?: string | null;
  weather: Weather;
});